+++ static/js/editor/modal-manager.js
@@ -410,3 +410,5 @@ if (typeof window !== 'undefined') {
 
 
 
+
+